refactor: 将数据库改为mysql及将主键id改为int64

+++ internal/account/model/account.go
@@ -1,8 +1,8 @@
 package model
 
 type Account struct {
-	Id          int     `db:"id" json:"id,omitempty"`
-	UserId      int     `db:"user_id" json:"user_id,omitempty"`
+	Id          int64   `db:"id" json:"id,omitempty"`
+	UserId      int64   `db:"user_id" json:"user_id,omitempty"`
 	AccountType *int    `db:"account_type" json:"account_type" binding:"required,numeric"`
 	AccountName string  `db:"account_name" json:"account_name" binding:"required,max=30"`
 	Balance     float64 `db:"balance" json:"balance" binding:"required,numeric"`
